+++ web/bookmarks/style.css
@@ -1,1 +1,1 @@
-@import '../components/bookmark.css'
+@import '../components/bookmark/bookmark-list.css'
