+++ web/components/bookmark/bookmark-edit.css
@@ -1,0 +1,29 @@
+.bc-bookmark-edit
+input[type="date"],
+input[type="datetime-local"],
+input[type="email"],
+input[type="month"],
+input[type="number"],
+input[type="password"],
+input[type="search"],
+input[type="tel"],
+input[type="text"],
+input[type="time"],
+input[type="url"],
+input[type="week"] {
+  width: 100%;
+}
+
+.bc-bookmark-note {
+  display: block;
+  width: 100%;
+  resize: vertical;
+}
+
+.bc-bookmark-legend {
+  font-size: var(--small-font-size);
+}
+
+.bc-bookmark-url-edit {
+  font-family: var(--font-code);
+}
