+++ web/components/bookmark/bookmark-list.css
@@ -1,0 +1,2 @@
+@import './bookmark-view.css';
+@import './bookmark-edit.css';
