+++ web/components/header/index.js
@@ -1,0 +1,1 @@
+index.js
