+++ web/components/star/index.css
