+++ web/components/star/index.js
